requirements.txt
tqdm
pdfminer.six
googletrans
nltk
